LayoutTests/imported/w3c/web-platform-tests/2dcontext/drawing-rectangles-to-the-canvas/2d.clearRect.globalalpha-expected.txt

2d.clearRect.globalalpha
clearRect is not affected by globalAlpha
Actual output:
Expected output:


PASS clearRect is not affected by globalAlpha 
